feat: refactor attribute creation methods for testing and consolidate functionality
